+++ src/Quotelinks/QuoteYou.js
@@ -6,9 +6,9 @@ import Header from "../General/Header";
 import { Conf, d, doc, g } from "../globals/globals";
 import Menu from "../Menu/Menu";
 import ExpandComment from "../Miscellaneous/ExpandComment";
+import ScrollMarkers from "../Miscellaneous/ScrollMarkers";
 import $ from "../platform/$";
 import $$ from "../platform/$$";
-import { debounce } from "../platform/helpers";
 import PostRedirect from "../Posting/PostRedirect";
 
 /*
@@ -57,17 +57,6 @@ var QuoteYou = {
     });
 
     QuoteYou.menu.init();
-
-    QuoteYou.scrollMarkerContainer = $.el('div', { classList: 'scroll-marker-container' });
-    doc.insertAdjacentElement('afterbegin', QuoteYou.scrollMarkerContainer);
-    $.on( QuoteYou.scrollMarkerContainer, 'click', (e) => {
-      const { postId } = /** @type {HTMLElement} */ (e.target).dataset;
-      if (postId) Header.scrollTo(g.posts[postId].nodes.root);
-    })
-    // $.on(d, 'PostsInserted', QuoteYou.markScroll);
-    // $.on(d, 'PostsRemoved', QuoteYou.markScroll);
-    $.on(window, 'resize', QuoteYou.markScroll);
-    new ResizeObserver(QuoteYou.markScroll).observe(doc)
   },
 
   isYou(post) {
@@ -83,7 +72,7 @@ var QuoteYou = {
 
     if (QuoteYou.isYou(this)) {
       $.addClass(this.nodes.root, 'yourPost');
-      QuoteYou.markScroll();
+      ScrollMarkers.markScroll();
     }
 
     // Stop there if there's no quotes in that post.
@@ -140,38 +129,9 @@ var QuoteYou = {
           quoter.classList.toggle('quotesYou', !!$('.quotelink.you', quoter));
         }
       }
-      QuoteYou.markScroll();
+      ScrollMarkers.markScroll();
     }
   },
-
-  markScroll: debounce(100, () => {
-    // Remove previous
-    QuoteYou.scrollMarkerContainer.innerText = '';
-
-    g.posts?.forEach((post) => {
-      const postEl = post.nodes.root;
-      let isReply = false;
-      if ($.hasClass(postEl, 'quotesYou')) {
-        isReply = true;
-      } else if (!$.hasClass(postEl, 'yourPost')) {
-        return;
-      }
-
-      const postPosition = postEl.getBoundingClientRect();
-      const top = (((postPosition.top + window.scrollY) / doc.scrollHeight) * 100).toFixed(1);
-      const height = Math.max(1, (postPosition.height / doc.scrollHeight) * 100).toFixed(1);
-
-      const marker = $.el('div', {
-        classList: `post-scroll-marker ${isReply ? 'reply' : 'you'}-scroll-marker`,
-        ariaHidden: true,
-      });
-      marker.style.top = `${top}vh`;
-      marker.style.height = `${height}vh`;
-      marker.dataset.postId = `${post.boardID}.${post.ID}`;
-
-      $.add(QuoteYou.scrollMarkerContainer, marker);
-    })
-  }, false),
 
   cb: {
     seek(type) {